+++ lib/whatsapp_sdk/resource/business_profile.rb
@@ -1,0 +1,15 @@
+# typed: strict
+# frozen_string_literal: true
+
+module WhatsappSdk
+  module Resource
+    class BusinessProfile
+      extend T::Sig
+
+      VERTICAL_VALUES = %w[
+        UNDEFINED OTHER AUTO BEAUTY APPAREL EDU ENTERTAIN EVENT_PLAN FINANCE GROCERY
+        GOVT HOTEL HEALTH NONPROFIT PROF_SERVICES RETAIL TRAVEL RESTAURANT NOT_A_BIZ
+      ].freeze
+    end
+  end
+end